Fix 300+ compiler warnings and errors, only 3033 to go!
Lots of unused parameters and variables
More than a few values set but never used
Lots of bad switch() statements that don't have a default or use all enum values
Lots of bad cases using values that are not part of the enum
Lots of sprintf without length check!
The actual errors are difficult to see among all the warning noise.